+++ CHANGELOG.md
@@ -7,10 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   `Link`: add `--lumx-link-visited-color` and `--lumx-link-text-decoration` CSS variable theming.
+
 ### Fixed
 
 -   `@lumx/react`:
     -   `ImageLightbox`: `useImageLightbox`'s `getTriggerProps` now always includes `aria-haspopup="dialog"` to warn screen reader users that activating the trigger opens a modal dialog.
+
+### Documentation
+
+-   Storybook `Link`: add theming story
 
 ## [4.22.0][] - 2026-09-03